requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.12.0.88
numpy==2.2.6
torch==2.7.1+cu118
torchvision==0.22.1+cu118
ultralytics==8.0.186
pillow==11.0.0
python-multipart==0.0.6
requests==2.31.0
pydantic==2.5.0
scikit-learn==1.3.2
matplotlib==3.8.2
tqdm==4.66.1